+++ packages/bottom-tabs/CHANGELOG.md
@@ -3,6 +3,66 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [6.4.0](https://github.com/react-navigation/react-navigation/compare/@react-navigation/bottom-tabs@6.3.2...@react-navigation/bottom-tabs@6.4.0) (2022-09-16)
+
+
+### Features
+
+* add freezeOnBlur prop  ([#10834](https://github.com/react-navigation/react-navigation/issues/10834)) ([e13b4d9](https://github.com/react-navigation/react-navigation/commit/e13b4d9341362512ba4bf921a17552f3be8735c1))
+
+
+
+
+
+## [6.3.3](https://github.com/react-navigation/react-navigation/compare/@react-navigation/bottom-tabs@6.3.2...@react-navigation/bottom-tabs@6.3.3) (2022-08-24)
+
+**Note:** Version bump only for package @react-navigation/bottom-tabs
+
+
+
+
+
+## [6.3.2](https://github.com/react-navigation/react-navigation/compare/@react-navigation/bottom-tabs@6.3.1...@react-navigation/bottom-tabs@6.3.2) (2022-07-05)
+
+
+### Bug Fixes
+
+* ensure same @types/react version in repo ([#10663](https://github.com/react-navigation/react-navigation/issues/10663)) ([e662465](https://github.com/react-navigation/react-navigation/commit/e6624653fbbd931158dbebd17142abf9637205b6)), closes [#10655](https://github.com/react-navigation/react-navigation/issues/10655)
+
+
+
+
+
+## [6.3.1](https://github.com/react-navigation/react-navigation/compare/@react-navigation/bottom-tabs@6.3.0...@react-navigation/bottom-tabs@6.3.1) (2022-04-01)
+
+**Note:** Version bump only for package @react-navigation/bottom-tabs
+
+
+
+
+
+# [6.3.0](https://github.com/react-navigation/react-navigation/compare/@react-navigation/bottom-tabs@6.2.0...@react-navigation/bottom-tabs@6.3.0) (2022-04-01)
+
+
+### Features
+
+* add an ID prop to navigators ([4e4935a](https://github.com/react-navigation/react-navigation/commit/4e4935ac2584bc1a00209609cc026fa73e12c10a))
+
+
+
+
+
+# [6.2.0](https://github.com/react-navigation/react-navigation/compare/@react-navigation/bottom-tabs@6.1.0...@react-navigation/bottom-tabs@6.2.0) (2022-02-02)
+
+
+### Features
+
+* add prop for new container ([#9772](https://github.com/react-navigation/react-navigation/issues/9772)) ([3fb2140](https://github.com/react-navigation/react-navigation/commit/3fb21409d6d0b66266c6d5eded2014ef2ebbda0a))
+
+
+
+
+
 # [6.1.0](https://github.com/react-navigation/react-navigation/compare/@react-navigation/bottom-tabs@6.0.9...@react-navigation/bottom-tabs@6.1.0) (2022-01-29)