+++ packages/devtools/CHANGELOG.md
@@ -3,6 +3,49 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.0.10](https://github.com/react-navigation/react-navigation/compare/@react-navigation/devtools@6.0.8...@react-navigation/devtools@6.0.10) (2022-09-16)
+
+**Note:** Version bump only for package @react-navigation/devtools
+
+
+
+
+
+## [6.0.9](https://github.com/react-navigation/react-navigation/compare/@react-navigation/devtools@6.0.8...@react-navigation/devtools@6.0.9) (2022-08-24)
+
+**Note:** Version bump only for package @react-navigation/devtools
+
+
+
+
+
+## [6.0.8](https://github.com/react-navigation/react-navigation/compare/@react-navigation/devtools@6.0.7...@react-navigation/devtools@6.0.8) (2022-07-05)
+
+
+### Bug Fixes
+
+* ensure same @types/react version in repo ([#10663](https://github.com/react-navigation/react-navigation/issues/10663)) ([e662465](https://github.com/react-navigation/react-navigation/commit/e6624653fbbd931158dbebd17142abf9637205b6)), closes [#10655](https://github.com/react-navigation/react-navigation/issues/10655)
+
+
+
+
+
+## [6.0.7](https://github.com/react-navigation/react-navigation/compare/@react-navigation/devtools@6.0.6...@react-navigation/devtools@6.0.7) (2022-04-01)
+
+**Note:** Version bump only for package @react-navigation/devtools
+
+
+
+
+
+## [6.0.6](https://github.com/react-navigation/react-navigation/compare/@react-navigation/devtools@6.0.5...@react-navigation/devtools@6.0.6) (2022-04-01)
+
+**Note:** Version bump only for package @react-navigation/devtools
+
+
+
+
+
 ## [6.0.5](https://github.com/react-navigation/react-navigation/compare/@react-navigation/devtools@6.0.4...@react-navigation/devtools@6.0.5) (2022-01-29)
 
 **Note:** Version bump only for package @react-navigation/devtools